Add an option to use a fork of ng-annotate
ng-annotate is no longer maintained[1], and hence fails when applied to
source code containing modern JavaScript constructs, like import and
export. In particular, this prevents usage with Webpack's ES6 modules
support (issue #17).

[1] https://github.com/olov/ng-annotate/issues/245

To work around it, allow the user to specify a fork of ng-annotate,
which adds support for new JS syntax.